qca-wifi-oss: Init DFS Puncturing SM in RCSA callback
Currently, when DFS Puncturing and RCSA feature is enabled and
we receive an RCSA / NOL IE from a connected STA, we add the
radar affected channels to NOL list, derive the corresponding
puncture pattern from the NOL IE and do a CSA restart to the
channel with the derived puncture pattern.

But, once the NOL period expires, we do not un-puncture the radar
affected channel. Thus, the AP continues to operate in a punctured
BW even after NOL expiry.

To fix this, initialize the DFS puncturing SM, when processing the
NOL IE from the RE STA.

The following changes are also made as part of this commit:

1] In a standalone Root AP, we ignore the radar event on a
   punctured subchannel. Similarly, add code changes to ignore NOL IE
   when RCSA is received on a punctured subchannel.
2] Do not allow initializing Puncturing SM on a RE which has STA vap.
   DFS puncturing is not a valid use case when there is a STA Vap and
   it is connected. This is because, the RE STA vap (and thus AP vaps)
   should follow the puncture bitmap of the Root.

Change-Id: I503c9318365148d042ece227778dd415a9b08537
(cherry picked from commit f9d6f2b842552c0a1ed4767c3983d67a425e14c4)