qcacmn: Treat AFC payload with no freq / chan objects as valid
When parsing the AFC payload received from the AFC application,
if the frequency / channel objects are not present in the AFC
response, then the payload is discarded as invalid.
But such a payload is actually valid. It indicates no channels
are available for SP operation.

To fix this, do not discard the AFC payload in the cfg80211
layer if the frequency or channel object NL80211 attributes
are not present.

Change-Id: I78759182230f37d08c176be420b622c854f896c5
CRs-Fixed: 3810799