qcacmn: Remove the dead code from wifi_pos component
If wifi_pos_parse_req() returns the non-zero value,
the caller function prints an error message and returns.
Else, the refcount of PSOC is taken and again checks if
the return code is non-zero value. Here, the second check is
dead code. Avoid this dead code by having a single check for
non-zero return value

Change-Id: I80138059808daedc6d24806cfd77c7de6e03fbdc
CRs-Fixed: 3870266