qcacmn: Notify VDEV repurpose status within serialization
Driver sends VDEV repurpose confirmation to FW during command's
serialization release memory and if any command is in pending queue
that gets active before this release memory is called.
The VDEV flag of is link switch in progress gets reset post FW notify
and if the next command in serialization is disconnect it won't send
the RSO stop as the link switch in progress is still set.

To synchronize these two events, send the VDEV repurpose completion
to FW in the serialization lock before next command gets active.

Change-Id: Ic7ee84757b3eb610821e4870ae68199098422404
CRs-Fixed: 3881849