qcacmn: Update  PSD power in the beacon in 0.5dBm
The power in the beacon is quantified in increments of 0.5 dBm.
Direct division by 100 followed by a multiplication by 2 could
resultin loss of precision due to truncation.
For instance:
The AFC server transmits a power value of 35.6.
This is sent as: 35.6×100=3560
Dividing first by 100, then multiplying by 2
results in:(3560/100)=(35×2)=70

However, the correct approach is to multiply by 2 before
dividing by 100:(3560×2)÷100=71

This yields 71, which corresponds to 35.5 dBm,
the nearest value to 35.6 dBm
given the beacon’s resolution of 0.5 dBm.

When the AFC server transmits a signal strength of 35.6 dBm,
the beacon should approximate this to 35.5 dBm.
Adjusting to 36 dBm would be incorrect and could breach
regulatory standards. Therefore, it’s recommended to incorporate
the formula Floor(35.6 * 2) into the TPE to ensure compliance
and accuracy.
CRs-Fixed: 3826409
Change-Id: Iaebc369652f3ddef9cae4f8ed62a5d959508cbc7