qca-wifi: Retrigger new ACS request if already running
ACS can be triggered as part of the ACS report command.
During this time if ACS is triggered from hostapd, and since its
already running as part of ACS report, the request from hostapd is
ignored.

After the scan completion of ACS report there is no notification
of the ACS status to hostapd, resulting the hostapd not
sending start AP and vap is not brought up.

So if any ACS is running while ACS request is received from hostapd,
we cancel the ACS request and retrigger ACS, so on completion
notifiaction is sent to hostapd and the vap is brought up.

Change-Id: If71f2dee8ace9e173d3ce8a3f6e8a078f6f3ac2d
CRs-Fixed: 3878931
(cherry picked from commit 1d195aa22d01354df4552b6176f409e5f5da1cb4)