qcacmn: Optimization changes to improve iperf KPI
The ieee80211_is_chan_radar API is invoked in the datapath flow
to restrict transmission (TX) on radar-infected channels. This API
fetches all the 20 MHz channels of the current channel, queries
the regulatory authority to check if it is radar-infected, and
then returns an output. This operation impacts the throughput
numbers (KPI). To optimize the execution time of this API, write
an API mlme_dfs_getnol_status that fetches the number of NOL
channels in the radio. If the count is non-zero, then iterate
through all the sub-channels to find which one is radar-infected.
This optimization has been shown to increase KPI numbers in
internal testing.

Hence, add the API dfs_get_radar_status and its dispatcher API.

CRs-Fixed: 3807186
Change-Id: I1fad35253e8134380933c192a7397a7101db22fe