qcacmn: Duplicate beacon received on adjacent channel on 6 GHz
Handled duplicate beacons received on adjacent channel on 6 GHz.
AP that is transmitting Beacon frame as non-HT DUP PPDU, will be
transmitting the Beacon across the BW on the AP. Case where duplicate
beacon is received on a different channel can have fluctuation in
RSSI values.
Check RSSI reported in the scan result and mark mismatch if RSSI
reported is lower than RSSI stored in scan entry for the same BSS.
This is done considering this scan result is because of leaked signal.

Change-Id: I7d81774a95e88ce31e73e20d2e401623e29ecb11
CRs-Fixed: 3913586