qcacmn: Add an API to find the first active vdev
WMI_VDEV_ADFS_CH_CFG_CMDID is sent on the first vdev of the pdev.
However, the vdev can be in the DOWN/INIT/STOP state and FW expects
WMI_VDEV_ADFS_CH_CFG_CMDID must be sent on a VDEV which is active
(VDEV START sent). Hence, add an API to find out the first active
vdev of the pdev and send WMI_VDEV_ADFS_CH_CFG_CMDID on that vdev.
Also, remove the substate variable in wlan_vdev_chan_config_valid as
it is redundant and not used.

CRs-Fixed: 3923584
Change-Id: Ia54058b0328f9f688f8513b477e01c8a2b7fee79